[Design] #65 로딩 화면 뷰와 뷰컨으로 파일 분리
